Fix contract deploy.
