+++ contracts/utils/sign.py
@@ -3,7 +3,7 @@ import bitcoin
 from ethereum import utils
 from secp256k1 import PrivateKey
 from eth_utils import encode_hex
-from utils.utils import sha3
+from utils.utils import sol_sha3
 
 
 eth_prefix = "\x19Ethereum Signed Message:\n"
@@ -11,7 +11,7 @@ eth_prefix = "\x19Ethereum Signed Message:\n"
 
 def eth_privtoaddr(priv) -> str:
     pub = bitcoin.encode_pubkey(bitcoin.privtopub(priv), 'bin_electrum')
-    return "0x" + binascii.hexlify(sha3(pub)[12:]).decode("ascii")
+    return "0x" + binascii.hexlify(sol_sha3(pub)[12:]).decode("ascii")
 
 
 def eth_message_prefixed(msg: str) -> bytes:
@@ -21,7 +21,7 @@ def eth_message_prefixed(msg: str) -> bytes:
 def eth_message_hex(msg: str) -> bytes:
     msg = eth_message_prefixed(msg)
     msg_hex = encode_hex(msg)
-    return sha3(msg_hex)
+    return sol_sha3(msg_hex)
 
 
 def sign(data: str, private_key_seed_ascii: str):
